Modify dashboard. Merge all the previous VM into 1. Eliminate unwanted files -previous VM-.
