+++ app/src/main/AndroidManifest.xml
@@ -63,11 +63,12 @@
             android:exported="true" />
         <activity
             android:name=".framework.view.Dashboard"
-            android:exported="true" />
-        <intent-filter>
-            <action android:name="android.intent.action.MAIN" />
-            <category android:name="android.intent.category.LAUNCHER" />
-        </intent-filter>
+            android:exported="true" >
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+        </activity>
         <activity
             android:name=".framework.view.AddNewProjectForm"
             android:exported="true" />
