+++ app/src/main/res/layout/dashboard.xml
@@ -84,7 +84,7 @@
                     android:orientation="vertical">
 
                     <TextView
-                        android:id="@+id/eventCount"
+                        android:id="@+id/eventRevenue"
                         android:layout_width="match_parent"
                         android:layout_height="wrap_content"
                         android:layout_gravity="center"
@@ -151,7 +151,6 @@
                         android:layout_width="fill_parent"
                         android:layout_height="fill_parent"
                         android:layout_marginTop="5dp">
-
                     </com.github.mikephil.charting.charts.PieChart>
 
                 </LinearLayout>
